+++ django/applications/catmaid/tokens.py
@@ -1,0 +1,11 @@
+from django.contrib.auth.tokens import PasswordResetTokenGenerator
+
+
+class TokenGenerator(PasswordResetTokenGenerator):
+    """A simple user token generator for email confirmation.
+    """
+    def _make_hash_value(self, user, timestamp):
+        return f'{user.pk}{timestamp}{user.is_active}'
+
+
+account_activation_token = TokenGenerator()